requirements.txt
numpy
pillow
rasterio
requests
matplotlib
